requirements.txt
# Data visualization and analysis
matplotlib>=3.8.0
numpy>=1.23.2,<2.0
pandas>=2.0.0
seaborn>=0.12.0
pillow>=10.0.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0

# Machine Learning Dependencies
tensorflow>=2.15.0
tensorflow-metal>=1.0.0  # GPU acceleration for Apple Silicon
scikit-learn>=1.3.0

# Web Framework & UI
streamlit>=1.28.0

# Supporting packages
python-dateutil>=2.8.0
pytz>=2023.0